+++ app/assets/javascripts/crud/index.js
@@ -1,6 +1,7 @@
 //= require select2-full
 //= require select2_locale_ja
 //= require crud/select2
+//= require crud/modal_picker
 //= require bootstrap-datepicker
 //= require bootstrap-timepicker
 //= require bootstrap-filestyle