+++ app/inputs/select2_input.rb
@@ -9,13 +9,14 @@ class Select2Input < SimpleForm::Inputs::CollectionSelectInput
       if options[:selected_item].present?
         options[:collection] = [options[:selected_item]]
       else
-        options[:collection] = object.respond_to?(attribute_name) ? init_data(value) : []
+        options[:collection] = init_data(value)
       end
     end
 
     js = javascript_tag(<<-SCRIPT
       $(function() {
         $("##{input_id}").val(#{value.inspect}).crudSelect2(#{select2_options(input_options).to_json});
+        #{onChangeHintScript}
       });
       SCRIPT
     )
@@ -39,6 +40,32 @@ class Select2Input < SimpleForm::Inputs::CollectionSelectInput
     @multiple
   end
 
+  def onChangeHintScript
+    return @onChangeHintScript if @onChangeHintScript
+    if options[:onChangeHint]
+      @onChangeHintScript = <<-SCRIPT
+        var target = $("##{input_id}");
+        var hint = target.siblings("small.text-muted");
+        var hintBody, hintHtml = "";
+      SCRIPT
+      options[:collection].map do |resource|
+        @onChangeHintScript << <<-SCRIPT
+          hintBody = #{options[:onChangeHint].gsub(/\$data\.(\w+)/) {|property| resource.send($1).inspect}};
+          hintHtml += $("<span/>").html(hintBody).attr({"data-id": #{resource.id}, "class": "chnage_hint_effect"}).get(0).outerHTML;
+        SCRIPT
+      end
+      @onChangeHintScript << <<-SCRIPT
+        if (hint.length) {
+          hint.html(hintHtml);
+        } else {
+          var newHint = $("<small/>").attr("class", "form-text text-muted").html(hintHtml);
+          newHint.appendTo(target.parent());
+        }
+      SCRIPT
+      @onChangeHintScript
+    end
+  end
+
   def submit_event
     @submit_event ||= "submit"
   end
@@ -56,6 +83,7 @@ class Select2Input < SimpleForm::Inputs::CollectionSelectInput
   end
 
   def init_data(ids)
+    return [] if ids.blank?
     model ? model.where(id: ids) : ids
   end
 
@@ -67,6 +95,6 @@ class Select2Input < SimpleForm::Inputs::CollectionSelectInput
     name = reflection.try(:name) || attribute_name
     options[:placeholder] ||= I18n.t("simple_form.select2.placeholder", name: object.class.human_attribute_name(name))
     options[:language] = I18n.locale
-    options.except(:as, :collection).transform_keys {|key| key.to_s.camelize(:lower)}
+    options.except(:as, :collection, :model).transform_keys {|key| key.to_s.camelize(:lower)}
   end
 end