+++ spec/lib/crud/model_reflection_spec.rb
@@ -82,7 +82,7 @@ describe Crud::ModelReflection do
 
       it "#column_metadata" do
         expect(subject.column_metadata(:string)).to eq(name: :string, type: :string)
-        expect(subject.column_metadata(:enumerized)).to eq(name: :enumerized, type: :enum)
+        expect(subject.column_metadata(:enumerized)).to eq(name: :enumerized, type: :enum, original_type: :string)
         expect(subject.column_metadata(:zzz)).to be nil
       end